% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Splits the LOCATION \code{"COUNTRY: REGION"} column into COUNTRY and REGION
columns and eliminates it}
\usage{
eq_location_clean(data = NULL)
}
\arguments{
\item{data}{a dataframe or tibble with a LOCATION column}
}
\value{
dataframe
}
\description{
Splits the LOCATION \code{"COUNTRY: REGION"} column into COUNTRY and REGION
columns and eliminates it
}
\examples{

\dontrun{

# Retrieve all earthquake data for Mexico since 2000
data <- eq_read_data() \%>\%
           filter(COUNTRY == "MEXICO", YEAR >= 2000) \%>\%
           dplyr::filter(!is.na(TOTAL_DEATHS)) \%>\%
           dplyr::select(DATE, YEAR, COUNTRY, REGION, LONGITUDE, LATITUDE, MAG, TOTAL_DEATHS) \%>\%
           tidyr::drop_na()
}

}
