% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_read_data}
\alias{eq_read_data}
\title{Reads in tab-separated value (TSV) raw data taken from the NOAA database
Renames Year, Mo, Dy, Hr, Mn, Sec, Tsu, Vol, and "Location Name" for clarity
Converts column names to uppercase
Drops columns irrelevant to this project}
\usage{
eq_read_data(
  filename = system.file("extdata", "earthquakes.tsv", package = "NOAAEarthquakes")
)
}
\arguments{
\item{filename:}{A character string containing the path and filename of the TSV file}
}
\value{
dataframe
}
\description{
Reads in tab-separated value (TSV) raw data taken from the NOAA database
Renames Year, Mo, Dy, Hr, Mn, Sec, Tsu, Vol, and "Location Name" for clarity
Converts column names to uppercase
Drops columns irrelevant to this project
}
\examples{

\dontrun{

# Retrieve earthquake data from 2017
data <- eq_read_data() \%>\%
           dplyr::filter(YEAR == 2017) \%>\%
           dplyr::select(DATE, YEAR, COUNTRY, REGION, LONGITUDE, LATITUDE, MAG, TOTAL_DEATHS) \%>\%
           tidyr::drop_na()
}

}
