% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_plot.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{\code{GeomTimeline} defines the ggplot2 geom.
Use the \code{geom_timeline} function to create these layers.
The aesthetic \code{x} determines the placement of the circles along the line
and is required.
The aesthetics \code{color}, \code{size}, and \code{alpha} define the color,
radius and transparency of the circle and are optional.
The aesthetic \code{y} can be used for making multiple stratified timelines
and is also optional.}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomTimeline
}
\value{
\code{Geom*} object for rendering timelines on plots
}
\description{
\code{GeomTimeline} defines the ggplot2 geom.
Use the \code{geom_timeline} function to create these layers.
The aesthetic \code{x} determines the placement of the circles along the line
and is required.
The aesthetics \code{color}, \code{size}, and \code{alpha} define the color,
radius and transparency of the circle and are optional.
The aesthetic \code{y} can be used for making multiple stratified timelines
and is also optional.
}
\examples{

\dontrun{
# All deadly earthquakes in JAPAN, CHINA and NEPAL since 2000

data <- eq_clean_data() \%>\%
     dplyr::filter(!is.na(TOTAL_DEATHS)) \%>\%
     dplyr::select(DATE, YEAR, COUNTRY, REGION, LONGITUDE, LATITUDE, MAG, TOTAL_DEATHS) \%>\%
     tidyr::drop_na() \%>\%
     dplyr::filter(YEAR > 1999, COUNTRY \%in\% c("JAPAN", "CHINA","NEPAL"))

# Plot Timeline

ggplot2::ggplot() +
    ggplot2::layer(geom = GeomTimeline,
                   mapping = aes(x = data$DATE,
                                 y = data$COUNTRY,
                                 size  = data$MAG,
                                 color = data$TOTAL_DEATHS),
                   data        = data,
                   stat        = 'identity',
                   position    = 'identity',
                   show.legend = NA,
                   inherit.aes = TRUE,
                   params = list(na.rm = FALSE))
}

}
\keyword{datasets}
