% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_plot.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{\code{GeomTimelineLabel} defines the ggplot2 geom which draws line markers and text.
Use \code{\link{geom_timeline_label}} to add a GeomTimelineLabel layer to ggplot.
The aesthetic \code{x} specifies where to place a marker and is required.
The aesthetic \code{mag} determines what quantity is used from the input data
to determine which observations are marked and is required.
The aesthetic \code{label} determines the column used in the input data for labels.
The aesthetics \code{y}, \code{color}, and \code{alpha} determines the stratification,
color, and transparency and are optional.
The aesthetic \code{n_max} determines how many markers to place and is optional.}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomTimelineLabel
}
\value{
\code{Geom*} object for marking and labeling timelines on plots
}
\description{
\code{GeomTimelineLabel} defines the ggplot2 geom which draws line markers and text.
Use \code{\link{geom_timeline_label}} to add a GeomTimelineLabel layer to ggplot.
The aesthetic \code{x} specifies where to place a marker and is required.
The aesthetic \code{mag} determines what quantity is used from the input data
to determine which observations are marked and is required.
The aesthetic \code{label} determines the column used in the input data for labels.
The aesthetics \code{y}, \code{color}, and \code{alpha} determines the stratification,
color, and transparency and are optional.
The aesthetic \code{n_max} determines how many markers to place and is optional.
}
\examples{

\dontrun{

# All deadly earthquakes in JAPAN, CHINA and NEPAL since 2000
data <- eq_clean_data() \%>\%
     dplyr::filter(!is.na(TOTAL_DEATHS)) \%>\%
     dplyr::select(DATE, YEAR, COUNTRY, REGION, LONGITUDE, LATITUDE, MAG, TOTAL_DEATHS) \%>\%
     tidyr::drop_na()

    ggplot2::ggplot() +

    ggplot2::layer(geom = GeomTimelineLabel,
                   mapping = aes(x = data$DATE,
                                 y = data$COUNTRY,
                                 size = data$MAG,
                                 color = data$TOTAL_DEATHS),
                   data = data,
                   stat = 'identity',
                   position = 'identity',
                   show.legend = NA,
                   inherit.aes = TRUE,
                   params = list(na.rm = FALSE))
}

}
\keyword{datasets}
